Refactor Z-order logic to support 128-bit temporal range
- Extract Z-order and hashing logic into a dedicated zorder module.
- Implement 128-bit Morton encoding to support timestamps beyond 2106.
- Update PostgreSQL return types for Z-order functions to NUMERIC.
- Update background worker and planner hooks to use the new zorder module.
- Standardize 128-bit verification in unit and regression tests.

+++ sql/spiral.sql
@@ -170,7 +170,7 @@ $$ LANGUAGE plpgsql
 SET search_path = public, spiral, "$user";
 
 -- Missing functions from demo.sql
-CREATE OR REPLACE FUNCTION to_spiraling_number(t BIGINT, cycle INTEGER, lane INTEGER) RETURNS BIGINT AS $$
+CREATE OR REPLACE FUNCTION to_spiraling_number(t BIGINT, cycle INTEGER, lane INTEGER) RETURNS NUMERIC AS $$
     SELECT spiral_zorder($1, ARRAY[$2::text, $3::text]);
 $$ LANGUAGE SQL IMMUTABLE PARALLEL SAFE;
